move test to appropriate file
